+++ src/exceptions/exception.ts
